Refactored component to use translateText functions
