+++ src/components/TranslatedText/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
-import { useContext } from 'react'
-import { TranslationsContext } from '../../hooks/TranslationsContext'
+import { TranslateString } from '../../utils/translateText'
 
 export interface TranslatedTextProps {
   translationId: number
@@ -8,21 +7,6 @@ export interface TranslatedTextProps {
 }
 
 const TranslatedText = ({ translationId, children }: TranslatedTextProps) => {
-  const getTranslation = (translations: Array<any>, id: number) => {
-    const foundTranslation = translations.find(translation => {
-      return translation.data.stringId === id
-    })
-    return foundTranslation && foundTranslation.data.text
-  }
-
-  const TranslateString = (translationId: number, fallback: string) => {
-    const { translations } = useContext(TranslationsContext)
-    if (translations.length > 0) {
-      return getTranslation(translations, translationId)
-    }
-    return fallback
-  }
-
   return <>{TranslateString(translationId, children)}</>
 }
